% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{assess_multimodels}
\alias{assess_multimodels}
\title{Assess multiple pairs of measured and predicted values}
\usage{
assess_multimodels(data, ..., .metrics = c("simplerspec", "yardstick"),
  .model_name = "model")
}
\arguments{
\item{data}{Data frame with all measured (observed) and predicted variables.}

\item{...}{Multiple arguments with observed (measured)-predicted pairs,
specified with \code{dplyr::vars(o = <column_name>, p = <column_name>)}.
Column names can strings or symbols. The arguments in `...` need to be named.}

\item{.metrics}{Character vector with package used for metrics calculation.
Default is \code{"simplerspec"}, which uses
\code{simplerspec::evaluate_model()}.}

\item{.model_name}{String with name for the new column that specifies the
model or the outcome variable. Default is \code{"model"}.}
}
\value{
Data frame with with summary statistics for measured values and
performance metrics for the pairs of measured and predicted values.
}
\description{
Return performance metrics for test set predictions and
measured values, e.g. for different model outcome variables.
}
