% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spc-extended.R
\name{bind_lcols_dts}
\alias{bind_lcols_dts}
\title{Bind list-columns within a tibble into a list of data.tables}
\usage{
bind_lcols_dts(spc_tbl, lcols, spc_id = "unique_id", group_id = "sample_id")
}
\arguments{
\item{spc_tbl}{Spectral data in a tibble data frame (classes "tibble_df",
"tbl" and "data.frame").}

\item{lcols}{Character vector of column names of list-columns to be bound
into a list of data.tables}

\item{spc_id}{Character vector denoting column name for a unique spectrum ID.
Default is \code{"unique_id"}.}

\item{group_id}{Character vector denoting column name for the spectrum group
ID. Default is \code{"sample_id"}. The group ID can later be used for
plotting spectra by group (e.g. by using different colors or panels).}
}
\value{
A list of data.tables. Elements contain data from list-columns
specified in \code{lcols} argument as data.tables. All data.tables contain in
addition \code{spc_id} and \code{group_id} columns.
}
\description{
Bind one to many list-columns in spectral tibble into a list
of data.tables.
}
