% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-modeling.R
\name{control_train}
\alias{control_train}
\title{Perform model tuning}
\usage{
control_train(x, response, env = parent.frame())
}
\arguments{
\item{x}{list from calibration sampling}

\item{response}{response variable for PLS regression, supplied
as character expression}

\item{env}{Environment where function is evaluated}

\item{validation}{Logical expression weather an independent
validation is performed.}
}
\description{
Uses function from caret to to model tuning
for PLS regression.
}
