% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-vip.R
\name{create_vip_rects}
\alias{create_vip_rects}
\title{Create a data frame containing start and end positions (wavenumbers)
where variable importance in projection (VIP) is > 1}
\usage{
create_vip_rects(df_vip)
}
\arguments{
\item{df_vip}{Data frame containing \code{wavenumber} and \code{vip} columns
(numeric)}
}
\value{
Data.frame containing vectors \code{start} (numeric; wavenumber),
\code{end} (numeric; wavenumber) and group (integer; values are
\code{1:length(start))}.
}
\description{
Given a data frame with VIP outputs (wavenumber and vip
columns), start and end values denoting spectral regions where VIP > 1
are returned as data frame. The functions can be used as helper
function for plotting VIP.
}
