% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretreat-spectra.R
\name{do_pretreatment}
\alias{do_pretreatment}
\title{Preprocess spectra}
\usage{
do_pretreatment(list_spectra, select)
}
\arguments{
\item{list_spectra}{List that contains averaged spectra
in the list element called \code{MIR_mean}}

\item{select}{Character string that specifies the predefined
pretreatment options. Possible arguments are:
\code{select = "MIR0"} for Savitzky Golay smoothing filter
without derivative, \code{select = "MIR1"} for Savitky Golay
with first derivative, \code{select = "MIR2"} for Savitzky
Golay with second derivative, \code{select = "MIR0_snv"}
for Standard Normal Variate after Savitzky Golay without
derivative, and \code{select = "MIRb"} for
baseline correction.}
}
\value{
list_spectra: List that contains preprocessed
spectra in element \code{MIR0}
}
\description{
Use commonly used preprocessing algorithms on
the spectra.
}

