% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{evaluate_model}
\alias{evaluate_model}
\alias{summary_df}
\title{Calculate model evaluation metrics}
\usage{
evaluate_model(data, obs, pred)

summary_df(df, x, y)
}
\arguments{
\item{data}{\code{data.frame} with predicted and observed data in columns.}

\item{obs}{Column that contains observed values, \code{symbol}/\code{name} or
\code{character} (wrapped in "").}

\item{pred}{Column that contains predicted values, \code{symbol}/\code{name} or
\code{character} (wrapped in "").}

\item{df}{\code{data.frame} with predicted and observed data in columns.}

\item{x}{Column that contains observed values, \code{symbol}/\code{name} or
\code{character} (wrapped in "").}

\item{y}{Column that contains predicted values, \code{symbol}/\code{name} or
\code{character} (wrapped in "").}
}
\description{
Calculate summary statistics of observed values and model
evaluation statistics for assessing agreement between observed (\code{obs}) and
predicted (\code{pred}) values.
}
