% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-modeling.R
\name{evaluate_model_q}
\alias{evaluate_model_q}
\title{Evaluate PLS performance}
\usage{
evaluate_model_q(x, model, response, evaluation_method, tuning_method,
  print = TRUE, env = parent.frame())
}
\arguments{
\item{x}{List that contains calibration and validation data
frame with combined spectral and chemical data}

\item{model}{List with PLS regression model output from
the caret package}

\item{response}{Response variable (e.g. chemical property) to be
modelled (needs to be non-quoted expression). \code{response}
needs to be a column name in the \code{validation} data.frame
(element of \code{x})}

\item{print}{Print observed vs. predicted for calibration
and validation. Default is \code{TRUE}.}

\item{env}{Specifiy the environment in which the function is
called. Default argument of \code{env} is
\code{parent.frame()}}

\item{validation}{Logical expression if independent validation
is performed (split data set into calibration set and
validation set)}
}
\description{
Calculate model performance indices based
on observed and predicted values of validation and calibration
set, and internal cross-validation
}
