% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-modeling.R
\name{evaluate_pls_q}
\alias{evaluate_pls_q}
\title{Evaluate PLS performance}
\usage{
evaluate_pls_q(x, pls_model, variable, validation = TRUE, print = TRUE,
  env = parent.frame())
}
\arguments{
\item{x}{List that contains calibration and validation data
frame with combined spectral and chemical data}

\item{pls_model}{List with PLS regression model output from
the caret package}

\item{variable}{Response variable (e.g. chemical property) to be
modelled (needs to be non-quoted expression). \code{variable}
needs to be a column name in the \code{validation} data.frame
(element of \code{x})}

\item{validation}{Logical expression if independent validation
is performed (split data set into calibration set and
validation set)}

\item{print}{Print observed vs. predicted for calibration
and validation. Default is \code{TRUE}.}

\item{env}{Specifiy the environment in which the function is
called. Default argument of \code{env} is
\code{parent.frame()}}
}
\description{
Calculate model performance indices based
on observed and predicted values of validation and calibration
set, and internal cross-validation
}

