% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-lcol-map-funprog.R
\name{extract_lcols2dts}
\alias{extract_lcols2dts}
\title{Extract multiple tibble list-columns and return data as list of
data.tables}
\usage{
extract_lcols2dts(spc_tbl, lcols)
}
\arguments{
\item{spc_tbl}{Spectral tibble (data frame) with spectral data contained
in list-columns}

\item{lcols}{Character vector containing names of list-columns to be
extracted into a list of data.tables}
}
\value{
List of data.tables. Each element is a data.table derivied from a
list-column specified in the \code{lcols} argument.
}
\description{
Extract multiple tibble list columns, row bind them separately
into single data tables and return a list of data.tables.
}
