% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-vip.R
\name{extract_pls_vip}
\alias{extract_pls_vip}
\title{Extract VIPs (variable importance in the projection) for a PLS
regression model output returned from model fitting with
\code{simplerspec::fit_pls()}}
\usage{
extract_pls_vip(mout)
}
\arguments{
\item{mout}{Model output list returned from \code{simplerspec::fit_pls()}.}
}
\value{
A tibble data frame with columns \code{wavenumber} and correponding
VIP values in the column \code{vip} for the finally chosen PLS regression
model at the final number of PLS components.
}
\description{
VIPs are extracted based on the \code{finalModel} sublist
in the \code{caret::train} output contained in the \code{model} element
of the \code{simplerspec::fit_pls()} model output list. The VIPs for
derived number of PLS components in the \code{finalModel} are computed.
}
