% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-modeling.R
\name{fit_pls}
\alias{fit_pls}
\title{Fit a PLS regression model}
\usage{
fit_pls(x, validation = TRUE, response, env = parent.frame())
}
\arguments{
\item{x}{List that contains calibration
set, validation set, and model tuning options}

\item{validation}{Logical expression weather independent
validation is performed}

\item{response}{Response variable to be modeled}

\item{env}{Environment where function is evaluated}
}
\description{
Uses the caret package to perform PLS modeling.
Spectra are centered and scaled prior to modeling.
}
