% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-modeling.R
\name{fit_rf_q}
\alias{fit_rf_q}
\title{Fit a random forest model
(quoted version of the function)}
\usage{
fit_rf_q(x, validation = TRUE, variable, tr_control, ntree_max = 500,
  env = parent.frame())
}
\arguments{
\item{x}{List that contains calibration
set, validation set, and model tuning options}

\item{validation}{Logical expression weather independent
validation is performed}

\item{variable}{Response variable to be modeled}

\item{tr_control}{Object that defines controlling parameters
of the desired internal validation framework}

\item{env}{Environment where function is evaluated}
}
\description{
Uses the caret package to perform random forest
modeling.
Spectra are centered and scaled prior to modeling.
}
