% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-chem-spectra.R
\name{join_chem_spec}
\alias{join_chem_spec}
\title{Join chemical and spectral data frames}
\usage{
join_chem_spec(dat_chem, dat_spec, by = "sample_ID")
}
\arguments{
\item{dat_chem}{data.frame that contains chemical values of
the sample}

\item{dat_spec}{List that contains spectral data}

\item{by}{character of column name that defines sample_ID}
}
\value{
List: xxx
}
\description{
Combines spectral data (data.frame) and chemical
data (data.frame).
}
