% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-chem-spectra.R
\name{join_spc_chem}
\alias{join_spc_chem}
\title{Join spectra data and chemical data tibbles}
\usage{
join_spc_chem(spc_tbl, chem_tbl, by = "sample_id")
}
\arguments{
\item{spc_tbl}{Tibble that contains spectral data}

\item{chem_tbl}{Tibble that contains chemical reference values of
the samples}

\item{by}{character of column name that defines sample_ID}
}
\value{
Tibble joined by sample_id
}
\description{
Combines spectral data (tibble class) and chemical
data (tibble class).
}

