% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-modeling.R
\name{ken_stone_q}
\alias{ken_stone_q}
\title{Split spectra into calibration and validation sets}
\usage{
ken_stone(spec_chem, ratio_val, pc, print = TRUE,
validation = TRUE)
}
\arguments{
\item{spec_chem}{data.frame that contains chemical
and IR spectroscopy data}

\item{ratio_val}{Ratio of number of validation and all samples.}

\item{pc}{Number of principal components (numeric)}

\item{print}{logical expression weather calibration}

\item{validation}{Logical expression weather
calibration sampling is performed
(\code{TRUE} or \code{FALSE}).}
}
\description{
Perform calibration sampling based on
the Kennard-Stones algorithm.
}

