% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-vip.R
\name{plot_pls_vip}
\alias{plot_pls_vip}
\title{Plot stacked ggplot2 graphs with the Variable Importance for the
Projection (VIP) scores, mean replicate spectra (absorbance) per sample_id, and the preprocessed spectra.}
\usage{
plot_pls_vip(mout, y1 = "spc_mean", y2 = "spc_pre", by = "sample_id",
  xlab = expression(paste("Wavenumber [", cm^-1, "]")),
  ylabs = c("Absorbance", "Preprocessed Abs."), alpha = 0.2)
}
\description{
Plot stacked ggplot2 graphs of VIP for the final
PLS regression model of the calibration (training) data set for the final
number of components, raw (replicate mean) spectra, and preprocessed spectra.
}
