% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-modeling.R
\name{pls_ken_stone}
\alias{pls_ken_stone}
\title{Calibration sampling, model tuning, and PLS regression}
\usage{
pls_ken_stone(spec_chem, response, variable = NULL,
  evaluation_method = "test_set", validation = TRUE,
  split_method = "ken_stone", tuning_method = "resampling",
  resampling_method = "kfold_cv", cv = "kfold_cv", ratio_val = 1/3,
  invert = TRUE, pls_ncomp_max = 20, ncomp_fixed = 5, pc = 2,
  print = TRUE, center = TRUE, scale = TRUE, env = parent.frame())
}
\arguments{
\item{spec_chem}{data.frame that contains IR spectroscopy
and chemical data}

\item{response}{Response variable (without quotes)}

\item{validation}{Logical expression weather independent
validation is performed}

\item{ratio_val}{Ratio of number of validation and all samples.}

\item{pc}{Number of Principal Components used for Calibration
sampling (Kennard-Stones algorithm)}

\item{print}{Logical expression weather graphs shall be printed}

\item{env}{Environment where function is evaluated}

\item{k}{Number of validation samples}
}
\description{
Perform calibration sampling and use selected
calibration set for model tuning
}
