% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-spc.R
\name{predict_from_spc}
\alias{predict_from_spc}
\title{Predict soil properties of new spectra based on a list of calibration models}
\usage{
predict_from_spc(spc_tbl, model_list, slice = TRUE)
}
\arguments{
\item{model_list}{List of model output generated from calibration step
(\code{pls_ken_stone()}}

\item{spc_tbl}{Tibble of spectra after preprocessing
(\code{preprocess_spc()})}

\item{slice}{Logical expression wheather only one row per sample_id returned.}
}
\value{
tibble with new columns \code{model}, and predicted values with
column names of model list.
}
\description{
Append predictions for a set of responses specified by a list
of calibration models and a tibble containing preprocessed spectra as
list-columns
}
