% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-spectra.R
\name{predict_from_spectra}
\alias{predict_from_spectra}
\title{Predict soil properties of new spectra based on calibration models}
\usage{
predict_from_spectra(model_list, spectra_list)
}
\arguments{
\item{model_list}{List that contains caret output objects
of the different calibration models to predict (one model per soil property)}

\item{spectra_list}{List that contains spectra and additional data
after pre-processing (\code{do_pretreatment()}including metadata
(\code{sample_ID})}
}
\description{
Function that uses pre-processed spectra, additional metadata of new
samples, and caret model output for the different soil property models
to create predicted values.
}
