% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-opus-universal.R
\name{read_opus_univ}
\alias{read_opus_univ}
\title{Read a list of Bruker OPUS spectrum binary files.}
\usage{
read_opus_univ(fnames, extract = c("spc"), parallel = FALSE)
}
\arguments{
\item{fnames}{List of character vectors containing full path names of spectra}

\item{extract}{Character vector of spectra types to extract from file.
Possible values are: "spc" (AB block in Bruker Opus software), "spc_nocomp"
(Spectra before final atmospheric compensation; only present if background
correction has been set in Opus), "ScSm" (Single channel spectrum of the
sample), "ScRf" (Single channel spectrum of the sample), "IgSm" (Interferogram
of the sample), "IgRf" (Interferogram of the reference). Default is
 \code{extract = c("spc")}.}

\item{parallel}{Logical (\code{TRUE} or \code{FALSE} indicating whether
files are read in parallel (multiple processors or multiple cores)).
Default is \code{parallel = FALSE}. If \code{TRUE} a parallel backend needs
to be registered, e.g. by using the \code{doParallel} package.}
}
\value{
out List spectra and metadata (parameters) extracted from
Bruker OPUS spectrometer files. List names are the names of the OPUS
files whose spectral data were extracted.
}
\description{
Read multiple spectral files measured with a Bruker FTIR Instrument. Files
containing spectra are in OPUS binary format.
\code{read_opus_univ} is a wrapper for \code{read_opus_bin_univ()})
}
