% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-spectra.R
\name{readOPUS_bin}
\alias{readOPUS_bin}
\title{Read an OPUS binary file}
\usage{
readOPUS_bin(file.name)
}
\arguments{
\item{file.name}{Character vector with path to files}
}
\description{
Read single binary file acquired with an
Bruker Vertex FTIR Instrument
}

