% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-spectra.R
\name{readOPUS_text}
\alias{readOPUS_text}
\title{Read an OPUS text file}
\usage{
readOPUS_text(file.name)
}
\arguments{
\item{file.name}{Character vector with path to files}
}
\description{
Read single text file acquired with
an Bruker Vertex FTIR Instrument
(as exported from OPUS software)
}

