% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample-spc.R
\name{resample_spc}
\alias{resample_spc}
\title{Resample spectra from spectral data object (tibble)}
\usage{
resample_spc(spc_tbl, x_unit = "wavenumber", wn_lower = 510,
  wn_upper = 3988, wn_interval = 2, wl_lower = 350, wl_upper = 2500,
  wl_interval = 1)
}
\arguments{
\item{spc_tbl}{Spectra data in a tibble object (classes "tibble_df", "tbl"
and "data.frame"). The spectra tibble is expected to contain at least
the columns 'spc' (list-column with spectral matrices stored in a list) and
`wavenumbers` or `wavelengths` (list-column that contains list of x-axis values
corresponding to each spectrum in `spc` as wavenumber or wavelength in
numeric values).}

\item{x_unit}{Character vector of the x-axis unit of the spectra in
within \code{spc_tbl}. Default is \code{"wavenumber"}. A further possible
argument is \code{"wavelength"}, which is usually the unit reported for
VIS/NIR spectrometers such as ASD field spectrometers.}

\item{wn_lower}{Numeric (integer or float) value of lowest wavenumber.
This argument will only be used if \code{x_unit = "wavenumber"}. The value
will be used as starting value when creating the new wavenumber sequence
that the spectra will be resampled upon. Default value is 510.}

\item{wn_upper}{Numeric (integer or float) value of highest wavenumber.
This argument will only be used if \code{x_unit = "wavenumber"}. The value
will be used as last value when creating the new wavenumber sequence
that the spectra will be resampled upon. Default value is 3988.}

\item{wn_interval}{Numeric value (integer or float) value of the wavenumber
increment when creating the new wavenumber sequence that the spectra will be
resampled upon. Default value is 2.}

\item{wl_lower}{Numeric (integer or float) value of lowest wavelength.
This argument will only be used if \code{x_unit = "wavelength"}. The value
will be used as starting value when creating the new wavenumber sequence
that the spectra will be resampled upon. Default value is 350.}

\item{wl_upper}{Numeric (integer or float) value of highest wavelength.
This argument will only be used if \code{x_unit = "wavelength"}. The value
will be used as last value when creating the new wavenumber sequence
that the spectra will be resampled upon. Default value is 2500.}

\item{wl_interval}{Numeric value (integer or float) value of the wavelength
increment when creating the new wavenumber sequence that the spectra will be
resampled upon. This argument will only be used if \code{x_unit = "wavelength"}.
Default value is 1.}
}
\value{
A spectra tibble containing the following additional columns
added to the input spectra tibble (`spc_tbl`): \code{wavenumbers_rs} or
\code{wavelengths_rs} as list-columns containing the resampled wavenumbers
or wavelengths, \code{spc_rs} list-column of resampled spectra returned
as list of data.tables (class "data.table" and "data.frame").
}
\description{
Resamples spectra stored in tibble column after
gathering spectra by \code{simplerspec::gather_spc()}.
}
