% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample-spectra.R
\name{resample_spectra}
\alias{resample_spectra}
\title{Resample spectra stored to new}
\usage{
resample_spectra(list_spectra, wn_lower = 510, wn_upper = 3988,
  wn_interval = 2)
}
\arguments{
\item{list_spectra}{List of spectra and metadata}

\item{wn_lower}{Numerical value for lowest  wavenumber in sampling interval}

\item{wn_upper}{Numerical value for highest wavenumber in sampling interval}
}

