% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-modeling.R
\name{rf_ken_stone}
\alias{rf_ken_stone}
\title{Calibration sampling, model tuning, and random forest modeling}
\usage{
rf_ken_stone(spec_chem, split_method = "ken_stone", ratio_val, pc = 2,
  print = TRUE, validation = TRUE, variable, ntree_max = 500,
  env = parent.frame())
}
\arguments{
\item{spec_chem}{data.frame that contains IR spectroscopy
and chemical data}

\item{ratio_val}{Ratio of number of validation and all samples.}

\item{pc}{Number of Principal Components used for Calibration
sampling (Kennard-Stones algorithm)}

\item{print}{Logical expression weather graphs shall be printed}

\item{validation}{Logical expression weather independent
validation is performed}

\item{variable}{Response variable (without quotes)}

\item{env}{Environment where function is evaluated}

\item{k}{Number of validation samples}
}
\description{
Perform calibration sampling and use selected
calibration set for model tuning
}

