% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-ref-spectra.R
\name{select_ref_samples}
\alias{select_ref_samples}
\title{Select a set of reference spectra to be measured by reference analysis
methods}
\usage{
ken_stone(spec_chem, ratio_val, pc, print = TRUE,
validation = TRUE)
}
\arguments{
\item{list_spectra}{list that contains preprocessed spectra and metadata}

\item{ratio_ref}{Ratio of desired reference samples to total sample number}

\item{pc}{Number of principal components (numeric). If pc < 1, the number
of principal components kept corresponds to the number of components
explaining at least (pc * 100) percent of the total variance.}

\item{print}{logical expression whether a plot (ggplot2) of sample selection
for reference analysis is shown in PCA space}

\item{validation}{Logical expression whether
calibration sampling is performed
(\code{TRUE} or \code{FALSE}).}
}
\description{
Select a set of calibration spectra to develop spectral models.
Samples in this list will be analyzed using laboratory reference methods.
}
