% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-spc.R
\name{select_spc_vars}
\alias{select_spc_vars}
\title{Select every n-th spectral variable for all spectra and x-values in spectral
tibble (\code{spc_tbl})}
\usage{
select_spc_vars(
  spc_tbl,
  lcol_spc = "spc_pre",
  lcol_xvalues = "xvalues_pre",
  every = NULL
)
}
\arguments{
\item{spc_tbl}{Tibble data.frame containing spectra in list-column}

\item{lcol_spc}{List-column containing spectra, specified with column
name as symbols or 1L character vector.}

\item{lcol_xvalues}{List-column containing x-values, specified with
column name as symbols or 1L character vector.}

\item{every}{Every n-th spectral positions to keep as 1L integer vector.}
}
\value{
a spectral tibble
}
\description{
Select every n-th spectral variable for all spectra and x-values in spectral
tibble (\code{spc_tbl})
}
