% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice-spc.R
\name{slice_xvalues}
\alias{slice_xvalues}
\title{Slice spectra into defined x-axis ranges}
\usage{
slice_xvalues(
  spc_tbl,
  xunit_lcol = "wavenumbers",
  spc_lcol = "spc",
  xvalues_cut = NULL
)
}
\arguments{
\item{spc_tbl}{Spectral data in a tibble object (classes "tibble_df", "tbl"
and "data.frame"). The spectra tibble is expected to contain at least
the column  \code{spc} (list-column with spectral matrices stored in a list)
and \code{wavenumbers} or \code{wavelengths} (list-column that contains list
of x-axis values).}

\item{xunit_lcol}{Character vector that specifies column name where x-axis
axis units are stored within \code{spc_tbl}. Default is \code{"wavenumber"}.}

\item{spc_lcol}{Character vector that specifies which column (list-column)
contains spectra to be sliced. Default is \code{"spc"}.}

\item{xvalues_cut}{List of numeric vectors that contains upper and lower bounds of respective regions to keep in spectra. The spectral regions outside
the \code{xvalues_cut} intervals will be cut out in the output spectra.}
}
\value{
Spectral tibble (data frame with list-columns) with sliced x-axis
column and spectral column. Both the x-axis list-column and the spectral
tibble list-column only contain data specified within the \code{xvalues_cut}
argument (list of numeric vectors).
}
\description{
Slice spectra contained in list-column of spectral tibble
(data frame). A list of x-axis value ranges can be specified. Spectra are
cut based on these ranges.
}
