% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soilspec_yamsys}
\alias{soilspec_yamsys}
\title{Soil spectra and laboratory reference data from Baumann et al. (2021)}
\format{
\subsection{\code{soilspec_yamsys}}{

A tibble data frame with 284 rows and 40 columns. The spectra are in the
\code{spc} list-column.
}
}
\source{
\url{https://soil.copernicus.org/articles/7/717/2021/}
}
\usage{
soilspec_yamsys
}
\description{
Data from "Estimation of soil properties with mid-infrared soil spectroscopy
across yam production landscapes in West Africa".
}
\keyword{datasets}
