% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-lcol-map-funprog.R
\name{split_df2l}
\alias{split_df2l}
\title{Split a tibble data frame into a list of tibbles by a group column}
\usage{
split_df2l(tbl_df, group)
}
\arguments{
\item{tbl_df}{Tibble data frame}

\item{group}{Character vector with name of column based on which tibble
is split into a list of tibbles}
}
\value{
List of tibbles. Each tibble contains data split by
a group column within \code{tbl_df}.
}
\description{
Helper function that calls \code{split} on a tibble using a
grouping column within tibble.
}
