% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra-utils.R
\name{summary_df}
\alias{summary_df}
\title{Calculate model evaluation metrics}
\usage{
summary_df(df, x, y)
}
\arguments{
\item{df}{data.frame with predicted and observed data}

\item{x}{column with observed values}

\item{y}{column with predicted values}
}
\description{
Calculates model statistics for predicted (y)
vs. observed (y) values
}
