% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-modeling.R
\name{train_pls}
\alias{train_pls}
\title{Fit a PLS regression model}
\usage{
train_pls(x, response, evaluation_method = "resampling",
  env = parent.frame())
}
\arguments{
\item{x}{List that contains calibration
set, validation set, and model tuning options}

\item{response}{Response variable to be modeled}

\item{env}{Environment where function is evaluated}

\item{validation}{Logical expression weather independent
validation is performed}
}
\description{
Uses the caret package to perform PLS modeling.
Spectra are centered and scaled prior to modeling.
}
