% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-modeling.R
\name{tune_model_loocv_q}
\alias{tune_model_loocv_q}
\title{Perform model tuning}
\usage{
tune_model_loocv_q(x, variable, env = parent.frame(), validation = TRUE)
}
\arguments{
\item{x}{list from calibration sampling}

\item{variable}{response variable for PLS regression, supplied
as character expression}

\item{env}{Environment where function is evaluated}

\item{validation}{Logical expression weather an independent
validation is performed.}
}
\description{
Uses function from caret to to model tuning
for PLS regression.
}

