% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-modeling.R
\name{tune_model_q}
\alias{tune_model_q}
\title{Perform model tuning}
\usage{
tune_model_q(x, response, env = parent.frame(),
  evaluation_method = "resampling")
}
\arguments{
\item{x}{list from calibration sampling}

\item{response}{response variable for PLS regression, supplied
as character expression}

\item{env}{Environment where function is evaluated}

\item{validation}{Logical expression weather an independent
validation is performed.}
}
\description{
Uses function from caret to to model tuning
for PLS regression.
}
