% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average-spc.R
\name{average_spc}
\alias{average_spc}
\title{Average spectra in spectral tibble data frame}
\usage{
average_spc(spc_tbl, by = "sample_id")
}
\arguments{
\item{spc_tbl}{Tibble data.frame containing spectra in list-column
\code{spc_rs}. This list-column is created when resampling spectra
with \code{resample_spc()}}

\item{by}{Character vector of length 1; specifies column by which spectra
are averaged. Default is \code{"sample_id"}.}
}
\value{
Tibble data frame (class `tbl_df`) with mean spectra appended
as list-column named \code{spc_mean}.
}
\description{
Average spectra in tibble by a group column after
resampling spectra by \code{simplerspec::resample_spc()}.
}
