% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average-spc.R
\name{average_spc}
\alias{average_spc}
\title{Average spectra from spectral data object (tibble)}
\usage{
average_spc(spc_tbl)
}
\description{
Averages spectra in tibble column by sample_id after
resampling spectra by \code{simplerspec::resample_spc()}.
}
