% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Calculate model evaluation metrics}
\usage{
evaluate_model(data, obs, pred)
}
\arguments{
\item{data}{`data.frame`` with predicted and observed data in columns.}

\item{obs}{Column that contains observed values, `symbol`/`name` or
`character` (wrapped in "").}

\item{pred}{Column that contains predicted values, `symbol`/`name` or
`character` (wrapped in "").}
}
\description{
Calculate observed summary statistics and model evaluation
statistics for assessing agreement between observed (`obs`) and
predicted (`pred`) values.
}
