% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather-spc.R
\name{gather_spc}
\alias{gather_spc}
\title{Gather spectra from list of spectral data into a tibble object}
\usage{
gather_spc(data)
}
\arguments{
\item{data}{list with file name elements that contain spectra and metadata
after reading binary OPUS files with \code{simplerspec::read_opus()}}
}
\value{
Spectral data and metadata in object of class tible
}

