% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spc.R
\name{plot_spc}
\alias{plot_spc}
\title{Plot tibble spectra}
\usage{
plot_spc(spc_tbl = pls_$data$calibration,
spc_tbl_2 = pls_C$data$validation,
y = "spc_pre", by = "sample_id", graph_id_1 = "Calibration",
graph_id_2 = "Validation", slice = TRUE, alpha = 1)
}
\description{
Plot spectra from tibble spectra objects.
}
