% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-spc.R
\name{preprocess_spc}
\alias{preprocess_spc}
\title{Preprocess spectra from spectral data object (tibble)}
\usage{
preprocess_spc(spc_tbl, select)
}
\description{
Preprocesses spectra in tibble column by sample_id after
averaging spectra by \code{simplerspec::average_spc()}.
}

