% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-asd.R
\name{read_asd}
\alias{read_asd}
\title{Read tab delimited text (.txt) files from ASD field spectrometer data
export into simplerspec spectra tibble.}
\usage{
read_asd(file)
}
\value{
Spectra data in tibble data frame (class `tbl_df`) that contains
columns `sample_id`, `spc` (list of spectral matrices) and 'wavelengths'
(list of wavelength vectors).
}
\description{
For reading ASD file spectrometer data the exported data are
expected to be in one .txt tab delimited file. The first row should contain
the name 'Wavelength' for the first column and the file names for the
remaining columns.
}
