% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-opus-universal.R
\name{read_opus_bin_univ}
\alias{read_opus_bin_univ}
\title{Read a Bruker OPUS spectrum binary file}
\usage{
read_opus_bin_univ(file_path)
}
\arguments{
\item{file_path}{Character vector with path to file}
}
\description{
Read single binary file acquired with an
Bruker Vertex FTIR Instrument
}
