% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-opus-universal.R
\name{read_opus_bin_univ}
\alias{read_opus_bin_univ}
\title{Read a Bruker OPUS spectrum binary file}
\usage{
read_opus_bin_univ(file_path, extract = c("spc"),
print_progress = TRUE, atm_comp_minus4offset = FALSE)
}
\arguments{
\item{file_path}{Character vector with path to file}

\item{extract}{Character vector of spectra types to extract from OPUS binary
file. Default is \code{c("spc")}, which will extract the final spectra, e.g.
expressed in absorbance (named \code{AB} in Bruker OPUS programs). Possible
additional values for the character vector supplied to extract are
\code{"ScSm"} (single channel spectrum of the sample measurement), \
code{"ScRf"} (single channel spectrum of the reference measurment),
\code{"IgSm"} (interferogram of the sample measurment) and \code{"IgRf"}
(interferogram of the reference measurement).}

\item{print_progress}{Logical (default \code{TRUE}) whether a message is
printed when an OPUS binary file is parsed into an R list entry.}

\item{atm_comp_minus4offset}{Logical whether spectra after atmospheric
compensation are read with an offset of \code{-4} bites from Bruker OPUS
files. Default is \code{FALSE}.}
}
\description{
Read single binary file acquired with an
Bruker Vertex FTIR Instrument
}
