% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample-spc.R
\name{resample_spc}
\alias{resample_spc}
\title{Resample spectra from spectral data object (tibble)}
\usage{
resample_spc(spc_tbl, wn_lower = 510, wn_upper = 3988, wn_interval = 2)
}
\description{
Resamples spectra stored in tibble column after
gathering spectra by \code{simplerspec::gather_spc()}.
}

