% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-spc.R
\name{select_spc_vars}
\alias{select_spc_vars}
\title{Select every n-th spectral variable for all spectra and x-values in spectral
tibble (`spc_tbl`)}
\usage{
select_spc_vars(spc_tbl, lcol_spc = spc_pre,
  lcol_xvalues = xvalues_pre, every = NULL)
}
\arguments{
\item{every}{}
}
\description{
Select every n-th spectral variable for all spectra and x-values in spectral
tibble (`spc_tbl`)
}
